gcp/website/pyproject.toml
[project]
name = "website-backend"
requires-python = ">=3.11,<4.0"
dependencies = [
    "Flask==2.3.3",
    "Flask-Caching==2.3.0",
    "Flask-Compress==1.17",
    "werkzeug==3.1.3",
    "google-auth==2.38.0",
    "google-cloud-ndb==2.3.2",
    "google-cloud-logging==3.11.3",
    "jinja2==3.1.5",
    "markdown2==2.5.3",
    "markupsafe==2.1.5",
    "packageurl-python==0.16.0",
    "redis==4.6.0",
    "requests==2.32.3",
    "gunicorn==22.0.0",
    "whitenoise==6.8.2",
    "cvss==3.3",
    "osv",
]

[tool.poetry]
package-mode = false 

[tool.poetry.dependencies]
osv = { path = "../../", develop = true }

[tool.poetry.group.dev.dependencies]
yapf = "*"
pylint = "*"


[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"